memory_accounting

Module allocator

Global allocator implementation that allows tracking allocations on a per-group basis.

Structs

AllocationGroupRegistry

A registry of allocation groups and the statistics for each of them.

AllocationGroupToken

A token associated with a specific allocation group.

AllocationStats

Statistics for an allocation group.

AllocationStatsSnapshot

Snapshot of allocation statistics for a group.

Tracked

An object wrapper that tracks allocations and attributes them to a specific group.

TrackingAllocator

A global allocator that tracks allocations on a per-group basis.

TrackingGuard

A guard representing an allocation group which has been entered.

Traits

Track

Attaches allocation groups to a Future.