memory_accounting

Struct MemoryBoundsBuilder

pub struct MemoryBoundsBuilder<'a> { /* private fields */ }

Builder for defining the memory bounds of a component and its subcomponents.

This builder provides a simple interface for defining the minimum and firm bounds of a component, as well as declaring subcomponents. For example, a topology can contain its own “self” memory bounds, and then define the individual bounds for each component in the topology.

Implementations

impl MemoryBoundsBuilder<'_>

pub fn reset(&mut self)

Resets the bounds of the current component to a default state.

This can be used in scenarios where the bounds of a component need to be redefined after they have been specified, as not all components are able to be defined in a single pass.

pub fn minimum(&mut self) -> BoundsBuilder<'_, Minimum>

Gets a builder object for defining the minimum bounds of the current component.

pub fn firm(&mut self) -> BoundsBuilder<'_, Firm>

Gets a builder object for defining the firm bounds of the current component.

The firm limit is additive with the minimum required memory, so entries that are added via minimum do not need to be added again here.

pub fn subcomponent<S>(&mut self, name: S) -> MemoryBoundsBuilder<'_>
where S: AsRef<str>,

Creates a nested subcomponent and gets a builder object for it.

This allows for defining the bounds of various subcomponents within a larger component, which are then rolled up into the calculated bounds for the parent component.

pub fn with_subcomponent<S, C>(&mut self, name: S, component: &C) -> &mut Self
where S: AsRef<str>, C: MemoryBounds,

Creates a nested subcomponent based on the given component.

This allows for defining a subcomponent whose bounds come from an object that implements MemoryBounds directly.